refactor: twofactor auth process
