--- front/src/UI/atoms/TwoFactorAuthInput/index.scss
@@ -1,13 +1,0 @@
-.opt_input {
-  height: 2rem;
-  margin: 1rem;
-  border-width: thin;
-  text-align: center;
-}
-
-.otp_input_container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  height: 100vh;
-}
