+++ front/src/UI/molecules/TwoFactorAuthInput/TwoFactorAuthInputQueries.ts
