+++ front/src/UI/molecules/TwoFactorAuthInput/index.scss
@@ -1,0 +1,6 @@
+.opt_input {
+  height: 2rem;
+  margin: 1rem;
+  border-width: thin;
+  text-align: center;
+}
