+++ front/src/UI/organisms/TwoFactorAuthBox/index.scss
@@ -1,0 +1,7 @@
+.two_factor_auth_container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: 100vh;
+}
